Refactor Submission Status
